Make bundler optional
For some projects I don't have bundler as part of the Gemfile but run it
as an external tool. With this kind of setup it would be nice to be able
to opt out of `bundle exec` even if you have a Gemfile and are using
bundler for the rest of the project.