feat(lsp): Add experimental use of lsp-format-buffer
Add support for using `lsp-format-buffer` to format Ruby buffers instead
of `rubocopfmt` when `rubocopfmt-mode` and `lsp-mode` are both enabled.

Disabled by default, and must manually be enabled by setting
`rubocopfmt-on-save-use-lsp-format-buffer` to `t`.

I consider this feature EXPERIMENTAL, meaning it "works on my machine",
with no guarantees of it working elsewhere, nor if it will remain part
of rubocopfmt in the future.